requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.34.2
openai==1.45.0
supabase==2.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
openpyxl==3.1.5
apscheduler==3.10.4
python-multipart==0.0.9
aiofiles==24.1.0
